+++ gameballsdk/src/main/java/com/gameball/gameball/model/request/RewardPointBody.java
@@ -2,7 +2,7 @@ package com.gameball.gameball.model.request;
 
 public class RewardPointBody extends PointTransactionParams
 {
-    public RewardPointBody(int amount, String transactionOnClientSystemId, String transactionKey)
+    public RewardPointBody(double amount, String transactionOnClientSystemId, String transactionKey)
     {
         super(amount,transactionKey,true);
         setAmount(amount);
